As of the resolution of https://bugzilla.mozilla.org/show_bug.cgi?id=297078 header names for the setRequestHeader method of the XMLHttpRequest object can no longer contain certain token characters. The colon is included in this list so I'm removing it from the user-agent header. This colon was a typo anyways.
As of the resolution of https://bugzilla.mozilla.org/show_bug.cgi?id=302263
the header 'content-length' is not be settable using on the XMLHttpRequest
object in mozilla browsers. I presume mozilla sets the content-length header
automatically. The XML-RPC spec requires the content-length header to be set
so I'll check to make sure other browsers do as well.


git-svn-id: file:///Users/gauthierm/svnrepo/xml-rpc-ajax/trunk@2126 94490950-a86c-4493-9ccb-530f72933e47